python/requirements.txt
google-cloud-storage
google-cloud-dialogflow-cx
